BBSDEV-29253 Handle split in OID
Bitbucket has moved to storing the LFS objects in S3 in the same way
they are on a hierarchical filesystem, on account of the new File Store
API being used. So now rather than LFS objects being stored with a key:
    git-lfs/<hierarchy>/<oid>
instead it is:
    git-lfs/<hierarchy>/<oid[0:1]>/<oid[2:63]>

This approach to limiting the number of files that could be in a
directory is not actually necessary with S3, since it is not a
hierarchical store. However Git LFS in Bitbucket has moved to using the
FileStore API/SPI abstraction, so the Git LFS plugin doesn't know if it
is storing in S3, some other object store, or a hierarchical filesystem
like ext4, etc. As such the Git LFS plugin splits the OID up so not all
LFS objects are stored in the same directory. This prevents performance
problems in the case a file store is used where this does actually
matter.